requirements.txt
# Core Framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Data Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
folium>=0.14.0

# Configuration
python-dotenv>=1.0.0
toml>=0.10.0

# Development
jupyter>=1.0.0
notebook>=6.5.0

# Utilities
openpyxl>=3.1.0
xlrd>=2.0.0
